# Run per-file conditions once