Pushing the per-file conditions from fixer functions up to a per-file pre-pass stage can save many function calls and repeat checks. This yields a small but notable performance improvement.

Benchmarked in two ways.

**First,** using [pytest-profiling](https://pypi.org/project/pytest-profiling/) to measure the test suite:

**Before:** 1574686 function calls (1538837 primitive calls) in 1.308 seconds
**After:** 1569930 function calls (1534037 primitive calls) in 1.297 seconds

A removal of 4756 function calls, which is pretty impressive given the test snippets are small. I don’t read much into the performance improvement here.

**Second,** running on a client project of 50,475 lines of code across 268 files. Each condition was run 10 times with hyperfine.

**Before:** 557.4 ms ± 5.9 ms

**After:** 546.2 ms ± 3.9 ms

About 2% faster.