@UnknownPlatypus 2024-02-12
oh so you did went for the conditional approach, nice!

I'm wondering if the cached_properties on the `State` class are still valuable after that?

Also, another idea I had was to leverage the fact that all these state conditions are mutually exclusive, having one of them return `True` should make every other ones return `False` without evaluating the regex at all.