@adamchainz 2024-02-12
> I'm wondering if the cached_properties on the `State` class are still valuable after that?

I think they are because not all fixers use the conditions. Some of them check other things first.

> Also, another idea I had was to leverage the fact that all these state conditions are mutually exclusive, having one of them return True should make every other ones return False without evaluating the regex at all.

I think that will be too minor an improvement to make it worth the extra code. It could also backfire as they’re heuristics, so making the first evaluated one trump the others could lead to unintended false negatives or even non-determinism if fixers run in different orders.