@UnknownPlatypus 2024-02-13
Just out of curiosity, what was the hyperfine command you used to benchmark your changes ?