@adamchainz 2024-05-10
Replying after some delay: I used `git ls-files -z -- '*.py' | xargs -0 django-upgrade --target-version 5.0`, per the readme. The project uses pre-commit but I wanted to cancel out the overhead of running `pre-commit`.