@UnknownPlatypus 2024-05-10
No worries! Was not very familiar with hyperfine at the time but did some research in between, it really is a nice tool!